refactor(catalog): move restore helpers into the requests layer
parseRestoreHeader / isEffectivelyArchived / RestoreStatus are generic S3
HEAD helpers with no React or Preview dependency, but they lived in
components/Preview/loaders/, so requestsUntyped.js imported UP into a
component — an inverted dependency. Move restore.ts (+ spec) next to
object.ts in containers/Bucket/requests/; Preview now imports DOWN into
requests, matching every other module.
