+++ catalog/app/containers/Bucket/PackageTree/PackageTree.tsx
@@ -14,7 +14,6 @@ import * as FileEditor from 'components/FileEditor'
 import Message from 'components/Message'
 import Placeholder from 'components/Placeholder'
 import * as Preview from 'components/Preview'
-import type { RestoreStatus } from 'components/Preview/loaders/restore'
 import * as Notifications from 'containers/Notifications'
 import cfg from 'constants/config'
 import type * as Routes from 'constants/routes'
@@ -36,6 +35,7 @@ import parseSearch from 'utils/parseSearch'
 import * as s3paths from 'utils/s3paths'
 import usePrevious from 'utils/usePrevious'
 import * as workflows from 'utils/workflows'
+import type { RestoreStatus } from '../requests/restore'
 
 import * as Download from '../Download'
 import { FileProperties } from '../FileProperties'
